% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{Import data}
\usage{
import_data(type = NULL)
}
\arguments{
\item{type}{default is null.}
}
\value{
A string.
}
\description{
This functions import database from duckdb
}
